+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [#1434](https://github.com/oshi/oshi/pull/1434): Document unsigned int MTU and update toString - [@dbwiddis](https://github.com/dbwiddis).
 * [#1440](https://github.com/oshi/oshi/pull/1440): Add ps backup for command line for macOS Big Sur compatibility - [@dbwiddis](https://github.com/dbwiddis).
 * [#1442](https://github.com/oshi/oshi/pull/1442), [#1443](https://github.com/oshi/oshi/pull/1443): FreeBSD CI; fix FreeBSD Test Failures - [@dbwiddis](https://github.com/dbwiddis).
+* [#1455](https://github.com/oshi/oshi/pull/1455): Fix hanging prstat call on Solaris thread details - [@dbwiddis](https://github.com/dbwiddis).
 * Your contribution here
 
 # 5.3.0 (2020-10-11), 5.3.1 (2020-10-18), 5.3.2 (2020-10-25), 5.3.3 (2020-10-28), 5.3.4 (2020-11-01), 5.3.5 (2020-11-11), 5.3.6 (2020-11-15)